split notify service to iface & impl
