+++ cloud/blockstore/apps/server/ya.make
@@ -11,7 +11,7 @@ PEERDIR(
     cloud/blockstore/libs/kms/iface
     cloud/blockstore/libs/kms/impl
     cloud/blockstore/libs/logbroker/iface
-    cloud/blockstore/libs/notify
+    cloud/blockstore/libs/notify/impl
     cloud/blockstore/libs/rdma/impl
     cloud/blockstore/libs/root_kms/impl
     cloud/blockstore/libs/service
