+++ cloud/blockstore/libs/notify/iface/public.h
@@ -12,7 +12,4 @@ using IServicePtr = std::shared_ptr<IService>;
 class TNotifyConfig;
 using TNotifyConfigPtr = std::shared_ptr<TNotifyConfig>;
 
-struct IJsonGenerator;
-using IJsonGeneratorPtr = std::unique_ptr<IJsonGenerator>;
-
 }   // namespace NCloud::NBlockStore::NNotify
