+++ cloud/blockstore/libs/notify/iface/ya.make
@@ -1,0 +1,20 @@
+LIBRARY()
+
+SRCS(
+    notify.cpp
+    config.cpp
+)
+
+PEERDIR(
+    cloud/blockstore/config
+    cloud/storage/core/libs/common
+    cloud/storage/core/libs/diagnostics
+
+    library/cpp/threading/future
+)
+
+END()
+
+RECURSE_FOR_TESTS(
+    ut
+)
