+++ cloud/blockstore/libs/notify/impl/ut/ya.make
@@ -1,4 +1,4 @@
-UNITTEST_FOR(cloud/blockstore/libs/notify)
+UNITTEST_FOR(cloud/blockstore/libs/notify/impl)
 
 INCLUDE(${ARCADIA_ROOT}/cloud/storage/core/tests/recipes/small.inc)
