+++ cloud/blockstore/libs/notify/ya.make
@@ -1,22 +1,4 @@
-LIBRARY()
-
-SRCS(
-    config.cpp
-    https.cpp
-    notify.cpp
-    json_generator.cpp
+RECURSE(
+    iface
+    impl
 )
-
-PEERDIR(
-    cloud/blockstore/libs/common
-    cloud/blockstore/libs/diagnostics
-
-    library/cpp/http/io
-    library/cpp/threading/future
-
-    contrib/libs/curl
-)
-
-END()
-
-RECURSE_FOR_TESTS(ut)
